+++ samplesheets/assayapps/microarray/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MicroarrayConfig(AppConfig):
+    name = 'samplesheets.assayapps.microarray'